requirements.txt
langchain
langchain-community
langchain-openai
unstructured
pdf2image
pytesseract
pillow
scrapy
streamlit
pandas
playwright
openai
python-dotenv
tqdm
pypdf==4.2.0
openpyxl==3.1.3
# scikit-learn==1.5.0 # Commenté temporairement pour résoudre un problème de build
matplotlib==3.9.0
pinecone
# sentence-transformers # Remplacé par l'API OpenAI
# torch # Remplacé par l'API OpenAI
langgraph
pymupdf
XlsxWriter
PyPDF2